fixed liquidity pool bug and wrote test for it

+++ src/ILiquidityPool.sol
@@ -24,7 +24,7 @@ interface ILiquidityPool {
         uint16 observationIndex,
         uint16 observationCardinality,
         uint16 observationCardinalityNext,
-        uint8 feeProtocol,
+        uint32 feeProtocol,
         bool unlocked
     );
 }
